pyproject.toml
[project]
name = "peer-review-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "bs4>=0.0.2",
    "chromadb>=1.3.5",
    "google-adk>=1.19.0",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.1.0",
    "langchain-text-splitters>=1.0.0",
    "mem0ai>=0.1.115",
    "python-dotenv>=1.2.1",
    "reportlab>=4.4.5",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.51.0",
    "nest-asyncio>=1.5.8",
]
